+++ sm-shop/src/main/java/com/salesmanager/shop/application/config/SocialConfig.java
@@ -17,8 +17,10 @@ import org.springframework.social.config.annotation.ConnectionFactoryConfigurer;
 import org.springframework.social.config.annotation.EnableSocial;
 import org.springframework.social.config.annotation.SocialConfigurer;
 import org.springframework.social.connect.ConnectionFactoryLocator;
+import org.springframework.social.connect.ConnectionRepository;
 import org.springframework.social.connect.UsersConnectionRepository;
 import org.springframework.social.connect.jdbc.JdbcUsersConnectionRepository;
+import org.springframework.social.connect.web.ConnectController;
 import org.springframework.social.facebook.connect.FacebookConnectionFactory;
 import org.springframework.social.facebook.security.FacebookAuthenticationService;
 import org.springframework.social.security.AuthenticationNameUserIdSource;
@@ -91,5 +93,12 @@ public class SocialConfig implements SocialConfigurer {
             dataSource, authenticationServiceLocator(), textEncryptor);
     conn.setTablePrefix(SchemaConstant.SALESMANAGER_SCHEMA + ".");
     return conn;
+  }
+
+  @Bean
+  public ConnectController connectController(
+      ConnectionFactoryLocator connectionFactoryLocator,
+      ConnectionRepository connectionRepository) {
+    return new ConnectController(connectionFactoryLocator, connectionRepository);
   }
 }
